feat(bill-parser): Phase 2.5 — refactor parser, add derogations + new entities detection
Refactor parser.ts (1533 lines) into 9 focused modules for maintainability.
Add extractDerogations() for repealing provisions (Gap 1 P0) — BOCG-14-A-116-1
now detects Ley 18/2011 derogation. Add extractNewEntities() for substantive
additions (Gap 3 P1) — detects registros, órganos, sistemas created by bills.
Add LLM false positive post-filter in analyze-bill.ts (Gap 4 P0). New DB tables
bill_derogations + bill_entities, exposed in API. 439 tests, 0 fail.
